only drop tables if they exist

+++ db/migrate/20170501040746_change_price_to_double.rb
@@ -1,6 +1,6 @@
 class ChangePriceToDouble < ActiveRecord::Migration[5.0]
   def change
-  	change_column(:carts, :price, :double)
-   	change_column(:cart_items, :price, :double) 
+  	change_column(:carts, :price, :decimal)
+   	change_column(:cart_items, :price, :decimal) 
   end
 end
